Clean template with Auth0 HTMX
