--- Endpoints/ApiEndpoints.cs
@@ -1,16 +1,0 @@
-﻿using BlazorMinimalApis.Endpoints.Api.Example;
-
-namespace BlazorMinimalApis.Endpoints;
-
-public static class ApiEndpoints
-{
-    public static WebApplication MapApiEndpoints(this WebApplication app)
-    {
-        var api = app.MapGroup("/api");
-
-        // maps to /api/example
-        api.MapGet("/example", ExampleHandler.Index);
-
-        return app;
-    }
-}
