--- Endpoints/Pages/Home/RandomNumber.razor
@@ -1,5 +1,0 @@
-﻿@code {
-	[Parameter] public int Num { get; set; }
-}
-
-<h4 class="text-orange-500 font-medium">Random Number is @Num</h4>
