+++ Pages/_Imports.razor
@@ -7,10 +7,6 @@
 @using Microsoft.JSInterop
 @using Microsoft.AspNetCore.Authorization
 @using BlazorMinimalApis
-@using BlazorMinimalApis.Endpoints.Pages
-@using BlazorMinimalApis.Endpoints.Pages.Shared
-@using BlazorMinimalApis.Endpoints.Pages.Shared.Layout
 @using BlazorMinimalApis.Lib
 @using BlazorMinimalApis.Lib.Views
-@using BlazorMinimalApis.Data
 @using BlazorMinimalApis.Lib.Helpers
