+++ appsettings.Development.json
@@ -6,8 +6,7 @@
     }
   },
   "Auth0": {
-    "Domain": "",
-    "ClientId": "",
-    "ClientSecret": ""
+    "Domain": "dev-bo8azkahrdj1dn2t.us.auth0.com",
+    "ClientId": "mW4zgKFXCHjfAulRIRIme71cylUYKqME",
   }
 }
